--- libs/data-structure/README.md
@@ -1,11 +1,0 @@
-# data-structure
-
-This library was generated with [Nx](https://nx.dev).
-
-## Building
-
-Run `nx build data-structure` to build the library.
-
-## Running unit tests
-
-Run `nx test data-structure` to execute the unit tests via [Jest](https://jestjs.io).
