--- libs/data-structure/project.json
@@ -1,19 +1,0 @@
-{
-  "name": "data-structure",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/data-structure/src",
-  "projectType": "library",
-  "tags": [],
-  "targets": {
-    "build": {
-      "executor": "@nx/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/libs/data-structure",
-        "main": "libs/data-structure/src/index.ts",
-        "tsConfig": "libs/data-structure/tsconfig.lib.json",
-        "assets": ["libs/data-structure/*.md"]
-      }
-    }
-  }
-}
