--- libs/data-structure/src/lib/Graph/Graph.types.ts
@@ -1,5 +1,0 @@
-import { Vertex } from "./Vertex";
-
-export type AdjacencyList = Map<number, Array<number>>;
-
-export type AdjacencyMatrix = number[][];
