--- libs/data-structure/src/lib/LinkedList/LinkedList.test.ts
@@ -1,9 +1,0 @@
-import { LinkedList } from './LinkedList'
-describe('LinkedList', ()=> {
-  it('initalize empty list', ()=> {
-    const list = new LinkedList();
-
-    expect(list).toBeInstanceOf(LinkedList);
-    expect(list.length).toBe(0);
-  })
-});
