+++ libs/data-structures/README.md
@@ -1,0 +1,11 @@
+# data-structures
+
+This library was generated with [Nx](https://nx.dev).
+
+## Building
+
+Run `nx build data-structures` to build the library.
+
+## Running unit tests
+
+Run `nx test data-structures` to execute the unit tests via [Jest](https://jestjs.io).
