+++ libs/data-structures/jest.config.ts
@@ -1,10 +1,10 @@
 export default {
-  displayName: 'data-structure',
+  displayName: 'data-structures',
   preset: '../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/libs/data-structure',
+  coverageDirectory: '../../coverage/libs/data-structures',
 };
