+++ libs/data-structures/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@bismarck/data-structure",
+  "name": "@bismarck/data-structures",
   "version": "0.0.1",
   "dependencies": {
     "tslib": "^2.3.0"
