+++ libs/data-structures/src/lib/Graph/Graph.types.ts
@@ -1,0 +1,5 @@
+import { Vertex } from './Vertex';
+
+export type AdjacencyList = Map<number, Array<number>>;
+
+export type AdjacencyMatrix = number[][];
