+++ libs/data-structures/src/lib/Graph/Vertex.ts
