+++ libs/data-structures/src/lib/LinkedList/LinkedList.test.ts
@@ -1,0 +1,9 @@
+import { LinkedList } from './LinkedList';
+describe('LinkedList', () => {
+  it('initalize empty list', () => {
+    const list = new LinkedList();
+
+    expect(list).toBeInstanceOf(LinkedList);
+    expect(list.length).toBe(0);
+  });
+});
