+++ libs/data-structures/src/lib/LinkedList/LinkedList.ts
