+++ libs/data-structures/src/lib/LinkedList/Node.ts
