+++ libs/data-structures/src/lib/LinkedList/index.ts
