+++ libs/data-structures/src/lib/Stack/Stack.ts
