+++ libs/data-structures/src/lib/Stack/index.ts
