+++ libs/data-structures/src/lib/data-structure.ts
